feat: Add Author entity
This commit adds the Author entity class, which represents authors in the database. The Author entity includes fields for the author's ID and name, as well as a many-to-many relationship with the Book entity. This allows books to have multiple authors and authors to be associated with multiple books.

The Author entity is annotated with the necessary annotations for mapping to the database, including @Entity, @Id, @GeneratedValue, and @ManyToMany. It also includes the necessary imports and annotations for Lombok.

This commit is related to the following user commits:
- Add annotation that enables Spring's asynchronous method execution capability, which allows the use of the @Async annotation on methods to run them asynchronously.
- Feat: Creating Author entity